requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
nidaqmx==0.9.0
matplotlib==3.8.2
websockets==12.0
python-multipart==0.0.6
pydantic-settings==2.1.0
